Feature/post comtroller

+++ app/Post.php
@@ -11,7 +11,7 @@ class Post extends Model
 
     use Sluggable;
 
-    protected $fillable = ['title', 'content'];
+    protected $fillable = ['title', 'content', 'date'];
 
     const IS_DRAFT = 0;
     const IS_PUBLIC = 1;
@@ -19,12 +19,12 @@ class Post extends Model
 
     public function category()
     {
-        return $this->hasOne(Category::class);
+        return $this->belongsTo(Category::class);
     }
 
     public function author()
     {
-        return $this->hasOne(User::class);
+        return $this->belongsTo(User::class, 'user_id');
     }
 
     public function tags()
@@ -70,8 +70,15 @@ class Post extends Model
 
     public function remove()
     {
-        Storage::delete('uploads/' . $this->image);
+        $this->removeImage();
         $this->delete();
+    }
+
+    public function removeImage()
+    {
+        if ($this->image != null) {
+            Storage::delete('uploads/' . $this->image);
+        }
     }
 
     public function uploadeImage($image)
@@ -79,24 +86,25 @@ class Post extends Model
         if ($image == null) {
             return;
         }
-        Storage::delete('uploads/' . $this->image);
+        $this->removeImage();
+
         $filename = str_random(10) . '.' . $image->extension();
-        $image->saveAs('uploads', $filename);
+        $image->storeAs('uploads', $filename);
         $this->image = $filename;
         $this->save();
     }
 
     public function getImage()
     {
-        if($this->image == null){
+        if ($this->image == null) {
             return '/img/no-user-image.png';
         }
-        return '/uploads' . $this->image;
+        return '/uploads/' . $this->image;
     }
 
     public function setCategory($id)
     {
-        if ($id) {
+        if (!$id) {
             return;
         }
 
@@ -127,13 +135,13 @@ class Post extends Model
 
     public function setFeatured()
     {
-        $this->is_feature = 1;
+        $this->is_featured = 1;
         $this->save();
     }
 
     public function setStandart()
     {
-        $this->is_feature = 0;
+        $this->is_featured = 0;
         $this->save();
     }
 
@@ -153,6 +161,23 @@ class Post extends Model
         }
 
         return $this->setFeatured();
+    }
+
+    public function getCategoryTitle()
+    {
+        if ($this->category != null) {
+            return $this->category->title;
+        }
+
+        return "Нет категории";
+    }
+
+    public function getTagsTitles()
+    {
+        if (!$this->tags->isEmpty()) {
+            return implode(" , ", $this->tags->pluck('title')->all());
+        }
+        return "Нет тегов";
     }
